requirements_custom.txt
action-tutorials-py==0.9.4
addict==2.4.0
ament-copyright==0.9.8
ament-cppcheck==0.9.8
ament-cpplint==0.9.8
ament-flake8==0.9.8
ament-index-python==1.1.0
ament-lint==0.9.8
ament-lint-cmake==0.9.8
ament-package==0.9.5
ament-pep257==0.9.8
ament-uncrustify==0.9.8
ament-xmllint==0.9.8
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
demo-nodes-py==0.9.4
domain-coordinator==0.9.2
examples-rclpy-executors==0.9.4
examples-rclpy-minimal-action-client==0.9.4
examples-rclpy-minimal-action-server==0.9.4
examples-rclpy-minimal-client==0.9.4
examples-rclpy-minimal-publisher==0.9.4
examples-rclpy-minimal-service==0.9.4
examples-rclpy-minimal-subscriber==0.9.4
filelock==3.9.0
fonttools==4.51.0
fsspec==2023.4.0
-e git+https://github.com/IDEA-Research/Grounded-Segment-Anything.git@4ddf65cf4b1bbe508680bdfb271d89a36ddcf35a#egg=groundingdino&subdirectory=GroundingDINO
huggingface-hub==0.22.2
idna==3.4
importlib-metadata==7.1.0
importlib-resources==6.4.0
Jinja2==3.1.2
kiwisolver==1.4.5
launch==0.10.10
launch-ros==0.11.7
launch-testing==0.10.10
launch-testing-ros==0.11.7
launch-xml==0.10.10
launch-yaml==0.10.10
MarkupSafe==2.1.3
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
osrf-pycommon==0.1.11
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pycocotools==2.0.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
quality-of-service-demo-py==0.9.4
regex==2024.4.16
requests==2.28.1
ros2action==0.9.13
ros2bag==0.3.11
ros2cli==0.9.13
ros2component==0.9.13
ros2doctor==0.9.13
ros2interface==0.9.13
ros2launch==0.11.7
ros2lifecycle==0.9.13
ros2multicast==0.9.13
ros2node==0.9.13
ros2param==0.9.13
ros2pkg==0.9.13
ros2run==0.9.13
ros2service==0.9.13
ros2topic==0.9.13
rosidl-runtime-py==0.9.1
rpyutils==0.2.0
rqt-action==1.0.2
rqt-console==1.1.2
rqt-graph==1.1.3
rqt-gui==1.1.2
rqt-gui-py==1.1.2
rqt-msg==1.1.1
rqt-plot==1.1.1
rqt-publisher==1.3.0
rqt-py-console==1.0.2
rqt-reconfigure==1.0.8
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-top==1.0.2
rqt-topic==1.3.0
safetensors==0.4.3
scipy==1.13.0
-e git+https://github.com/IDEA-Research/Grounded-Segment-Anything.git@4ddf65cf4b1bbe508680bdfb271d89a36ddcf35a#egg=segment_anything&subdirectory=segment_anything
six==1.16.0
sros2==0.9.5
supervision==0.19.0
sympy==1.12
teleop-twist-keyboard==2.3.2
timm==0.9.16
tokenizers==0.19.1
tomli==2.0.1
topic-monitor==0.9.4
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tqdm==4.66.2
transformers==4.40.0
triton==2.1.0
typing-extensions==4.8.0
urllib3==1.26.13
yapf==0.40.2
zipp==3.18.1
